+++ public/index.html
@@ -12,5 +12,7 @@
     />
     <title>Spotify Clone</title>
   </head>
-  <body id="root"></body>
+  <body>
+    <div id="root"></div>
+ </body>
 </html>
