+++ src/context.js
@@ -1,4 +1,4 @@
 import React from 'react';
 
 export const MainContext = React.createContext();
-export const HeaderContext = React.createContext();
+export const Context = React.createContext();
